On possible idle activity, duplicate data on a reactive flow.
We use a simple condition server-side to not roam on duplicates.  We
don't use SRTT/2 in that condition because even it seems a good idea,
and would probably works in practice, it is unsuitable on asymetric
links (with low latency on one direction and high latency on the
other).  Well, anyway, this will probably have no impact.