Simplify and change the sending procedure.
The procedure is now similar both client and server
side.  We send data on the fastest path, and compensate
loss by duplicating packets.  We may also send a
duplicate packet on the first reactive path, when all
the previous ones were idle.

Remark that now, the server does not necessarily use
the "last flow".  We expect that flow not used by the
client will naturally timeout.  This will probably
need to add RTO and idle calculation server-side.