Update way to sort flows server-side.
Old paths need to be discarded, and best flows selected. When 2 packets
are send simultaneously, selecting the path of the second packet means
selecting the slower path: we need to consider some idle time.  On the
other way, we can't use the client idle calculation, because the server
never waits for answer.